Update authoritations of films. Directors get details like films. LoginDTO requeriments added.
