+++ T1-APIREST/DTO/LoginDTO.cs
@@ -1,7 +1,11 @@
-﻿namespace T1_APIREST.DTO
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace T1_APIREST.DTO
 {
     public class LoginDTO
     {
+        [Required]
+        [EmailAddress]
         public string Email { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
